fix: jupyterlab_wipp-0.3.0
